Data Preparation Function

This function prepare the data for modelling.

data_preparation(
  data,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  outcome_cov = ~1,
  model_var = NULL,
  switch_n_cov = ~1,
  switch_d_cov = ~1,
  first_period = NA,
  last_period = NA,
  use_weight = 0,
  use_censor = 0,
  check_missing = 0,
  cense = NA,
  pool_cense = 0,
  cense_d_cov = ~1,
  cense_n_cov = ~1,
  include_followup_time_case = ~followup_time + I(followup_time^2),
  include_expansion_time_case = ~for_period + I(for_period^2),
  include_regime_length = FALSE,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  lag_p_nosw = 1,
  where_var = NULL,
  data_dir,
  save_weight_models = FALSE,
  chunk_size = 500,
  separate_files = FALSE,
  quiet = FALSE
)

Arguments

data

A data.frame containing all the required columns.

id

Name of the data column for id feature Defaults to id

period

Name of the data column for period feature Defaults to period

treatment

Name of the data column for treatment feature Defaults to treatment

outcome

Name of the data column for outcome feature Defaults to outcome

eligible

Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible

outcome_cov

A RHS formula with baseline covariates to adjust in final model

model_var

List of Variables of interest to be used in final model. Derived variables to use in outcome models. Typically assigned_treatment for ITT and per-protocol, and dose + dose^2 for as-treated. time_on_regime? TODO check what else is derived

switch_n_cov

A RHS formula for modelling probability of switching treatment. Used in the numerator of weight calculation.

switch_d_cov

A RHS formula for modelling probability of switching treatment. Used in the denominator of weight calculation.

first_period

First period value to start expanding about

last_period

Last period value to expand about

use_weight

Use weights in analysis. If 0 then no weights will be calculated.

use_censor

Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value

check_missing

Check for missing values in final model when use_censor=1 (Not added yet!)

cense

Censoring variable

pool_cense

Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model) Defaults to 0

cense_d_cov

A RHS formula for modelling probability of being censored. Used in the numerator of weight calculation.

cense_n_cov

A RHS formula for modelling probability of being censored. Used in the denominator of weight calculation.

include_followup_time_case

The model to include followup_time in outcome model, specified as a RHS formula.

include_expansion_time_case

The model to include for_period in outcome model, specified as a RHS formula.

include_regime_length

If defined as 1 a new variable (time_on_regime) is added to dataset. This variable stores the duration of time that the patient has been on the current treatment value

eligible_wts_0

Eligibility criteria used in weights for model condition Am1 = 0

eligible_wts_1

Eligibility criteria used in weights for model condition Am1 = 1

lag_p_nosw

When 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)

where_var

List of variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model

data_dir

Directory to model objects when save_weight_models=TRUE and expanded data as trial_i.csvs if separate_files = TRUE. If the specified directory does not exist it will be created. If the directory already contains trial files an error will occur, other files may be overwritten.

save_weight_models

Save weight models objects in data_dir.

chunk_size

Number of patients to process in one chunk when separate_files = TRUE

separate_files

Save expanded data in separate CSV files for each trial.

quiet

Don't print progress messages.

Details

The arguments chunk_size and separate_files allow for processing of large datasets that would not fit in memory once expanded. When separate_files = TRUE, the input data are processed in chunks of patients and saved into separate files for each trial starting period. These separate files can be sampled to create the dataset for the modelling.