Expand Function

This function performs the data expansion for a given dataset

expand(
  sw_data,
  outcomeCov_var,
  where_var,
  use_censor,
  followup_spline = NA,
  period_spline = NA,
  maxperiod,
  minperiod,
  lag_p_nosw,
  keeplist,
  data_dir,
  separate_files = FALSE
)

Arguments

sw_data

datatable to expand

outcomeCov_var

A list of individual baseline variables used in final model

where_var

Variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model

use_censor

Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value

followup_spline

The parameters for spline model for followup time when choose "spline" in the include_followup_time_case

period_spline

The parameters for spline model for for_period when choose "spline" in the include_expansion_time_case

maxperiod

Maximum period

minperiod

Minimum period

lag_p_nosw

when 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)

keeplist

A list contains names of variables used in final model

data_dir

Direction to save data

separate_files

Write to one file or one per trial (default FALSE)