INITIATORS Function

This function analyses observational longitudinal data to estimate the effect of interventions sustained over time. This emulates the design of a hypothetical randomized trial.

Usage,

initiators(
  data_path,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  outcomeCov_var = NA,
  outcomeCov = NA,
  outcomeClass = NA,
  model_var = NA,
  cov_switchn = NA,
  model_switchn = NA,
  class_switchn = NA,
  cov_switchd = NA,
  model_switchd = NA,
  class_switchd = NA,
  first_period = NA,
  last_period = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  run_unweighted_analysis = 0,
  run_weighted_analysis = 1,
  run_p99_analysis = 0,
  run_user_limits_analysis = 0,
  lower_weight = NA,
  upper_weight = NA,
  use_censor = 0,
  check_missing = 0,
  cense = NA,
  pool_cense = 0,
  cov_censed = NA,
  model_censed = NA,
  class_censed = NA,
  cov_censen = NA,
  model_censen = NA,
  class_censen = NA,
  include_followup_time_case = "linear",
  include_expansion_time_case = "linear",
  followup_spline = NA,
  period_spline = NA,
  include_regime_length = 0,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  lag_p_nosw = 1,
  where_var = NA,
  where_case = NA,
  run_base_model = 1,
  case_control = 0,
  n_control = 5,
  data_dir,
  numCores = NA,
  chunk_expansion = TRUE,
  chunk_size = 500,
  separate_files = FALSE,
  glm_function = "parglm"
)

Arguments

data_path

The path to csv file

id

Name of the data column for id feature Defaults to id

period

Name of the data column for period feature Defaults to period

treatment

Name of the data column for treatment feature Defaults to treatment

outcome

Name of the data column for outcome feature Defaults to outcome

eligible

Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible

outcomeCov_var

List of individual baseline variables used in final model

outcomeCov

List of functions of baseline covariates used in final model

outcomeClass

Categorical variables used in the final model

model_var

List of Variables of interest to be used in final model

cov_switchn

List of covariates to be used in logistic model for switching probabilities for numerator model

model_switchn

List of models (functions) to use the covariates from cov_switchn

class_switchn

Class variables used in logistic model for nominator model

cov_switchd

List of covariates to be used in logistic model for switching probabilities for denominator model

model_switchd

List of models (functions) to use the covariates from cov_switchd

class_switchd

Class variables used in logistic model for denominator model

first_period

First period value to start expanding about

last_period

Last period value to expand about

first_followup

First follow-up period

last_followup

Last follow-up period

use_weight

Use weights in analysis. If 0 then no weights will be calculated

run_unweighted_analysis

Run the final model with no weights when use_weights = 1

run_weighted_analysis

Run the final model with original weights

run_p99_analysis

Run the final model with truncating the weights at the 1st and 99th percentile

run_user_limits_analysis

Run the final model with truncating the weights using user defined limits

lower_weight

Use lower weight as minimum possible weight

upper_weight

Use upper weight as maximum possible weight

use_censor

Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value

check_missing

Check for missing values in final model when use_censor=1 (Not added yet!)

cense

Censoring variable

pool_cense

Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model) Defaults to 0

cov_censed

List of covariates to be used in logistic model for censoring weights in denominator model

model_censed

List of models (functions) to use the covariates from cov_censed

class_censed

Class variables used in censoring logistic regression in denominator model

cov_censen

List of covariates to be used in logistic model for censoring weights in numerator model

model_censen

List of models (functions) to use the covariates from cov_censen

class_censen

Class variables used in censoring logistic regression in numerator model

include_followup_time_case

The model to include follow up time in outcome model. This has 3 options c("linear","quadratic","spline")

include_expansion_time_case

The model to include for_period in outcome model. This has 3 options c("linear","quadratic","spline")

followup_spline

The parameters for spline model for followup time when choose "spline" in the include_followup_time_case (ex. list(df=2))

period_spline

The parameters for spline model for for_period when choose "spline" in the include_expansion_time_case (ex. list(df=3))

include_regime_length

If defined as 1 a new variable (time_on_regime) is added to dataset - This variable stores the duration of time that the patient has been on the current treatment value

eligible_wts_0

Eligibility criteria used in weights for model condition Am1 = 0

eligible_wts_1

Eligibility criteria used in weights for model condition Am1 = 1

lag_p_nosw

When 1 this will set the first weight to be 1 and use p_nosw_d and p_nosw_n at followup-time (t-1) for calculating the weights at followup-time t - can be set to 0 which will increase the maximum and variance of weights (Defaults to 1)

where_var

List of variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model

where_case

List of where conditions used in subsetting the data used in final analysis

run_base_model

Run the model with no conditions Defaults to 1

case_control

Run the case control sampling or not Defaults to 0

n_control

Number of controls used in case control sampling Defaults to 5

data_dir

Direction to save data

numCores

Number of cores for parallel programming (default value is maximum cores and parallel programming)

chunk_expansion

Do the expansion in chunks (and in parallel if numCores > 1). Turn this off if you have enough memory to expand the whole dataset at once. (default TRUE)

chunk_size

Number of ids to process at once for the chunk expansion (default 500). Larger chunk_sizes may be faster but require more memory.

separate_files

Write to one file or one per trial (default FALSE)

glm_function

Which glm function to use for the final model from stats or parglm packages

Details

The class variables paramers (outcomeClass,class_switchn,class_switchd,class_censen,class_censed) can be given as a character vector which will construct factors using as.factor or as a named list with the arguments for factor eg list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")

initiators()