Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
trial_example$catvarA <- factor(trial_example$catvarA)
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = c("catvarA", "nvarA"),
  use_censor = 0,
  use_weight = 0,
  last_period = 10
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.836   0.020   0.856 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   5.737   1.028   6.768 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 119036"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[, weight])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1266  -0.0965  -0.0858  -0.0771   3.6169  
#> 
#> Coefficients:
#>                      Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -6.506e+00  2.549e-01 -25.530  < 2e-16 ***
#> assigned_treatment  1.072e-01  1.844e-01   0.581 0.560972    
#> for_period          3.963e-03  1.950e-03   2.032 0.042159 *  
#> I(for_period^2)    -1.415e-06  4.817e-06  -0.294 0.768922    
#> followup_time       7.315e-03  2.157e-03   3.392 0.000693 ***
#> I(followup_time^2) -2.040e-05  8.647e-06  -2.359 0.018303 *  
#> catvarA1           -2.186e-01  1.291e-01  -1.692 0.090565 .  
#> catvarA2           -1.872e-01  1.687e-01  -1.110 0.267160    
#> catvarA3           -1.205e+01  1.822e+02  -0.066 0.947261    
#> catvarA7            1.022e-01  1.706e-01   0.599 0.549091    
#> nvarA              -2.509e-02  3.434e-02  -0.731 0.465071    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 5996.4  on 119035  degrees of freedom
#> Residual deviance: 5948.4  on 119025  degrees of freedom
#> AIC: 5970.4
#> 
#> Number of Fisher Scoring iterations: 16
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   1.306   0.288   1.609 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                  names      estimate    robust_se            lb            ub
#>  1:        (Intercept) -6.506479e+00 5.442370e-01 -7.573183e+00 -5.439774e+00
#>  2: assigned_treatment  1.072091e-01 3.219439e-01 -5.238010e-01  7.382192e-01
#>  3:         for_period  3.963214e-03 4.483238e-03 -4.823933e-03  1.275036e-02
#>  4:    I(for_period^2) -1.415271e-06 1.019418e-05 -2.139587e-05  1.856533e-05
#>  5:      followup_time  7.315365e-03 4.351074e-03 -1.212739e-03  1.584347e-02
#>  6: I(followup_time^2) -2.040151e-05 2.153166e-05 -6.260355e-05  2.180054e-05
#>  7:           catvarA1 -2.185521e-01 2.565036e-01 -7.212992e-01  2.841950e-01
#>  8:           catvarA2 -1.872242e-01 3.498992e-01 -8.730266e-01  4.985783e-01
#>  9:           catvarA3 -1.205101e+01 1.010951e+00 -1.403247e+01 -1.006954e+01
#> 10:           catvarA7  1.021860e-01 2.761929e-01 -4.391520e-01  6.435240e-01
#> 11:              nvarA -2.508872e-02 4.972880e-02 -1.225572e-01  7.237972e-02
#>               z  p_value
#>  1: -11.9552301  < 0.001
#>  2:   0.3330055 0.739130
#>  3:   0.8840069 0.376692
#>  4:  -0.1388312 0.889584
#>  5:   1.6812781 0.092709
#>  6:  -0.9475123 0.343378
#>  7:  -0.8520429 0.394190
#>  8:  -0.5350803 0.592594
#>  9: -11.9204634  < 0.001
#> 10:   0.3699807 0.711397
#> 11:  -0.5045110 0.613902
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period
#> (Intercept)         2.961939e-01      -2.744176e-03 -2.037192e-03
#> assigned_treatment -2.744176e-03       1.036479e-01  4.874758e-05
#> for_period         -2.037192e-03       4.874758e-05  2.009942e-05
#> I(for_period^2)     3.603084e-06      -2.823266e-07 -4.318695e-08
#> followup_time      -1.977478e-04      -8.433206e-05 -4.696819e-06
#> I(followup_time^2) -1.565598e-06       1.680108e-07  3.324697e-08
#> catvarA1           -1.861313e-02       2.911581e-03  3.752413e-05
#> catvarA2           -2.959740e-02       2.025207e-03  3.631664e-05
#> catvarA3           -2.900698e-02       6.911092e-03 -2.856804e-05
#> catvarA7           -6.884259e-02       5.211648e-03  3.816959e-04
#> nvarA               1.855334e-04      -8.541171e-05 -6.429315e-06
#>                    I(for_period^2) followup_time I(followup_time^2)
#> (Intercept)           3.603084e-06 -1.977478e-04      -1.565598e-06
#> assigned_treatment   -2.823266e-07 -8.433206e-05       1.680108e-07
#> for_period           -4.318695e-08 -4.696819e-06       3.324697e-08
#> I(for_period^2)       1.039214e-10  1.351413e-08      -7.370722e-11
#> followup_time         1.351413e-08  1.893184e-05      -8.402795e-08
#> I(followup_time^2)   -7.370722e-11 -8.402795e-08       4.636122e-10
#> catvarA1             -1.386524e-07 -3.085634e-05       1.907266e-07
#> catvarA2              2.493205e-08 -1.149529e-04       5.837749e-07
#> catvarA3              2.834734e-07 -1.605529e-04       9.451782e-07
#> catvarA7             -5.780150e-07 -1.867621e-04       1.133045e-06
#> nvarA                 6.939957e-09 -1.005276e-05      -1.663180e-08
#>                         catvarA1      catvarA2      catvarA3      catvarA7
#> (Intercept)        -1.861313e-02 -2.959740e-02 -2.900698e-02 -6.884259e-02
#> assigned_treatment  2.911581e-03  2.025207e-03  6.911092e-03  5.211648e-03
#> for_period          3.752413e-05  3.631664e-05 -2.856804e-05  3.816959e-04
#> I(for_period^2)    -1.386524e-07  2.493205e-08  2.834734e-07 -5.780150e-07
#> followup_time      -3.085634e-05 -1.149529e-04 -1.605529e-04 -1.867621e-04
#> I(followup_time^2)  1.907266e-07  5.837749e-07  9.451782e-07  1.133045e-06
#> catvarA1            6.579411e-02  2.826052e-02  1.766853e-02  1.658830e-02
#> catvarA2            2.826052e-02  1.224295e-01  2.640203e-02  2.939930e-02
#> catvarA3            1.766853e-02  2.640203e-02  1.022023e+00  2.538101e-02
#> catvarA7            1.658830e-02  2.939930e-02  2.538101e-02  7.628249e-02
#> nvarA              -1.412841e-03  1.084544e-03  8.620164e-04  7.062680e-04
#>                            nvarA
#> (Intercept)         1.855334e-04
#> assigned_treatment -8.541171e-05
#> for_period         -6.429315e-06
#> I(for_period^2)     6.939957e-09
#> followup_time      -1.005276e-05
#> I(followup_time^2) -1.663180e-08
#> catvarA1           -1.412841e-03
#> catvarA2            1.084544e-03
#> catvarA3            8.620164e-04
#> catvarA7            7.062680e-04
#> nvarA               2.472953e-03
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.144   0.004   0.147 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.003158780 0.006329388 0.009511663 0.012705438 0.015910464
#> 
#> $trt_1
#> [1] 0.003514863 0.007041489 0.010579676 0.014129223 0.017689823
#>