Read Data Function

This function read data from csv file and select the columns you need

read_data(
  data_address,
  data_path = NA,
  id_num = NA,
  id = "id",
  period = "period",
  treatment = "treatment",
  outcome = "outcome",
  eligible = "eligible",
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  outcomeCov_var = NA,
  cov_switchn = NA,
  cov_switchd = NA,
  cov_censed = NA,
  cov_censen = NA,
  cense = NA,
  where_var = NA
)

Arguments

data_address
Address for data read with bigmemory

data_path
Path of the csv file

id_num
Id number

id
Name of the data column for id feature Defaults to id

period
Name of the data column for period feature Defaults to period

treatment
Name of the data column for treatment feature Defaults to treatment

outcome
Name of the data column for outcome feature Defaults to outcome

eligible
Indicator of whether or not an observation is eligible to be expanded about Defaults to eligible

eligible_wts_0
Eligibility criteria used in weights for model condition Am1 = 0

eligible_wts_1
Eligibility criteria used in weights for model condition Am1 = 1

outcomeCov_var
List of individual baseline variables used in final model

cov_switchn
Covariates to be used in logistic model for switching probabilities for numerator model

cov_switchd
Covariates to be used in logistic model for switching probabilities for denominator model

cov_censed
Covariates to be used in logistic model for censoring weights for denominator model

cov_censen
Covariates to be used in logistic model for censoring weights for nominator model

cense
Censoring variable

where_var
Variables used in where conditions used in subsetting the data used in final analysis (where_case), the variables not included in the final model read_data()