Data modelling Function

This function do the modelling.

data_modelling(
  outcomeCov_var = NA,
  outcomeCov = NA,
  outcomeClass = NA,
  model_var = NA,
  first_followup = NA,
  last_followup = NA,
  use_weight = 0,
  run_unweighted_analysis = 0,
  run_weighted_analysis = 1,
  run_p99_analysis = 0,
  run_user_limits_analysis = 0,
  lower_weight = NA,
  upper_weight = NA,
  use_censor = 0,
  check_missing = 0,
  include_followup_time_case = c("linear", "quadratic", "spline"),
  include_expansion_time_case = c("linear", "quadratic", "spline"),
  where_case = NA,
  run_base_model = 1,
  absolutePath,
  numCores = NA,
  glm_function = c("parglm", "glm"),
  use_sample_weights = TRUE,
  quiet = FALSE
)

Arguments

outcomeCov_var

List of individual baseline variables used in final model

outcomeCov

List of functions of baseline covariates used in final model

outcomeClass

Categorical variables used in the final model

model_var

List of Variables of interest to be used in final model

first_followup

First follow-up period

last_followup

Last follow-up period

use_weight

Use weights in analysis. If 0 then no weights will be calculated

run_unweighted_analysis

Run the final model with no weights when use_weights = 1

run_weighted_analysis

Run the final model with original weights

run_p99_analysis

Run the final model with truncating the weights at the 1st and 99th percentile

run_user_limits_analysis

Run the final model with truncating the weights using user defined limits

lower_weight

Use lower weight as minimum possible weight

upper_weight

Use upper weight as maximum possible weight

use_censor

Use censoring for per-protocol analysis - censor person-times once a person-trial stops taking the initial treatment value

check_missing

Check for missing values in final model when use_censor=1 (Not added yet!)

include_followup_time_case

The model to include follow up time in outcome model. This has 3 options c("none","linear","quadratic","spline"). The data should contain the corresponding columns: "followup_time","followup_time2" or "followup_base".

include_expansion_time_case

The model to include for_period in outcome model. This has 3 options c("none","linear","quadratic","spline"). The data should contain the corresponding columns: "for_period","for_period2" or "period_base".

where_case

List of where conditions used in subsetting the data used in final analysis

run_base_model

Run the model with no conditions Defaults to 1

absolutePath

Direction to where the data for modelling is saved

numCores

Number of cores for parallel programming (default value is maximum cores and parallel programming)

glm_function

Which glm function to use for the final model from stats or parglm packages

use_sample_weights

Set model weights to sample_weight * weight.

quiet

Don't print progress messages. data_modelling()

Details

The class variables parameters (outcomeClass,class_switchn, class_switchd,class_censen,class_censed) can be given as a character vector which will construct factors using as.factor or as a named list with the arguments for factor e.g. list(risk_cat=list(levels = c(1,2,3,0), age_cat=list(levels=c(1,2,3),labels=c("50-60","60-70","70+")