Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
trial_example$catvarA <- factor(trial_example$catvarA)
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = c("catvarA", "nvarA"),
  use_censor = 0,
  use_weight = 0,
  last_period = 10
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.741   0.032   0.773 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   5.705   0.897   6.603 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 97039"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[, weight])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1385  -0.1033  -0.0911  -0.0788   3.4856  
#> 
#> Coefficients:
#>                      Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -6.481e+00  2.704e-01 -23.969  < 2e-16 ***
#> assigned_treatment -1.214e-01  2.400e-01  -0.506  0.61288    
#> for_period          6.246e-03  2.019e-03   3.094  0.00197 ** 
#> I(for_period^2)    -6.113e-06  4.963e-06  -1.232  0.21807    
#> followup_time       2.206e-03  2.050e-03   1.076  0.28171    
#> I(followup_time^2) -1.560e-06  7.525e-06  -0.207  0.83573    
#> catvarA1           -1.546e-02  1.330e-01  -0.116  0.90743    
#> catvarA2           -3.001e-01  1.941e-01  -1.547  0.12197    
#> catvarA3           -1.111e+01  2.308e+02  -0.048  0.96161    
#> catvarA7            3.396e-01  1.824e-01   1.862  0.06263 .  
#> nvarA              -3.408e-02  3.581e-02  -0.952  0.34125    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 5333.7  on 97038  degrees of freedom
#> Residual deviance: 5291.5  on 97028  degrees of freedom
#> AIC: 5313.5
#> 
#> Number of Fisher Scoring iterations: 15
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   0.932   0.208   1.161 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                  names      estimate    robust_se            lb            ub
#>  1:        (Intercept) -6.481253e+00 5.589004e-01 -7.576698e+00 -5.385808e+00
#>  2: assigned_treatment -1.214408e-01 3.311891e-01 -7.705715e-01  5.276898e-01
#>  3:         for_period  6.245769e-03 4.509880e-03 -2.593596e-03  1.508513e-02
#>  4:    I(for_period^2) -6.113261e-06 1.046666e-05 -2.662791e-05  1.440139e-05
#>  5:      followup_time  2.206441e-03 3.819833e-03 -5.280431e-03  9.693313e-03
#>  6: I(followup_time^2) -1.560285e-06 1.809706e-05 -3.703052e-05  3.390995e-05
#>  7:           catvarA1 -1.546137e-02 2.489305e-01 -5.033651e-01  4.724424e-01
#>  8:           catvarA2 -3.001220e-01 3.379106e-01 -9.624267e-01  3.621827e-01
#>  9:           catvarA3 -1.111058e+01 1.016455e+00 -1.310283e+01 -9.118324e+00
#> 10:           catvarA7  3.396072e-01 2.878848e-01 -2.246469e-01  9.038614e-01
#> 11:              nvarA -3.408422e-02 4.987472e-02 -1.318387e-01  6.367024e-02
#>                z p_value
#>  1: -11.59643733 < 0.001
#>  2:  -0.36668123 0.71386
#>  3:   1.38490794 0.16608
#>  4:  -0.58407007 0.55917
#>  5:   0.57762765 0.56352
#>  6:  -0.08621761 0.93129
#>  7:  -0.06211118 0.95047
#>  8:  -0.88816983 0.37445
#>  9: -10.93071239 < 0.001
#> 10:   1.17966387 0.23813
#> 11:  -0.68339670 0.49436
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period
#> (Intercept)         3.123696e-01      -1.294848e-02 -2.085810e-03
#> assigned_treatment -1.294848e-02       1.096862e-01  1.599483e-04
#> for_period         -2.085810e-03       1.599483e-04  2.033902e-05
#> I(for_period^2)     3.725766e-06      -6.538832e-07 -4.450599e-08
#> followup_time      -2.949571e-04      -4.672539e-05 -3.711701e-06
#> I(followup_time^2) -1.048580e-06       1.574548e-07  2.687573e-08
#> catvarA1           -3.109939e-02       1.086102e-02  4.857047e-05
#> catvarA2           -4.052448e-02       1.124149e-02  1.564211e-04
#> catvarA3           -3.253531e-02       9.826040e-03 -5.054944e-05
#> catvarA7           -8.334119e-02       1.184165e-02  4.684052e-04
#> nvarA              -8.672137e-04      -2.133697e-03 -5.967904e-06
#>                    I(for_period^2) followup_time I(followup_time^2)
#> (Intercept)           3.725766e-06 -2.949571e-04      -1.048580e-06
#> assigned_treatment   -6.538832e-07 -4.672539e-05       1.574548e-07
#> for_period           -4.450599e-08 -3.711701e-06       2.687573e-08
#> I(for_period^2)       1.095509e-10  1.223583e-08      -6.550288e-11
#> followup_time         1.223583e-08  1.459112e-05      -5.968456e-08
#> I(followup_time^2)   -6.550288e-11 -5.968456e-08       3.275034e-10
#> catvarA1             -1.090856e-07  6.799235e-05       2.381485e-08
#> catvarA2             -2.506535e-07 -2.171730e-04       1.230729e-06
#> catvarA3              3.357634e-07 -3.260312e-06       3.228905e-07
#> catvarA7             -8.137798e-07  1.648511e-05       1.724862e-07
#> nvarA                 8.918163e-09 -1.881221e-06      -1.844435e-08
#>                         catvarA1      catvarA2      catvarA3      catvarA7
#> (Intercept)        -3.109939e-02 -4.052448e-02 -3.253531e-02 -8.334119e-02
#> assigned_treatment  1.086102e-02  1.124149e-02  9.826040e-03  1.184165e-02
#> for_period          4.857047e-05  1.564211e-04 -5.054944e-05  4.684052e-04
#> I(for_period^2)    -1.090856e-07 -2.506535e-07  3.357634e-07 -8.137798e-07
#> followup_time       6.799235e-05 -2.171730e-04 -3.260312e-06  1.648511e-05
#> I(followup_time^2)  2.381485e-08  1.230729e-06  3.228905e-07  1.724862e-07
#> catvarA1            6.196639e-02  3.192954e-02  2.213874e-02  2.329351e-02
#> catvarA2            3.192954e-02  1.141835e-01  2.841550e-02  3.178294e-02
#> catvarA3            2.213874e-02  2.841550e-02  1.033180e+00  2.331396e-02
#> catvarA7            2.329351e-02  3.178294e-02  2.331396e-02  8.287763e-02
#> nvarA              -1.598848e-04 -1.181750e-03 -6.424394e-04 -3.886060e-04
#>                            nvarA
#> (Intercept)        -8.672137e-04
#> assigned_treatment -2.133697e-03
#> for_period         -5.967904e-06
#> I(for_period^2)     8.918163e-09
#> followup_time      -1.881221e-06
#> I(followup_time^2) -1.844435e-08
#> catvarA1           -1.598848e-04
#> catvarA2           -1.181750e-03
#> catvarA3           -6.424394e-04
#> catvarA7           -3.886060e-04
#> nvarA               2.487488e-03
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.131   0.000   0.132 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.004405700 0.008799655 0.013181844 0.017552247 0.021910796
#> 
#> $trt_1
#> [1] 0.003904044 0.007799844 0.011687374 0.015566609 0.019437486
#>