Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

# \donttest{
data("trial_example")
working_dir <- file.path(tempdir(), "trial_emu")
dir.create(working_dir)
data_path <- file.path(working_dir, "trial_example.csv")
write.csv(trial_example, file = data_path)

i <- initiators(
  data_path = data_path,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  model_var = "assigned_treatment",
  outcomeCov_var = c("catvarA", "catvarB", "catvarC", "nvarA", "nvarB", "nvarC"),
  outcomeClass = c("catvarA", "catvarB", "catvarC"),
  numCores = 1,
  data_dir = working_dir,
  use_censor = 0,
  use_weight = 0
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.783   0.052   0.836 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   6.145   0.999   7.146 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 1939053"
#> [1] "Base Analysis"
#> 
#> Call:
#> glm(formula = as.formula(regform), family = binomial(link = "logit"), 
#>     data = temp_data, weights = temp_data[, weight], control = parglm::parglm.control(nthreads = 4, 
#>         method = "FAST"), method = parglm::parglm.fit, singular.ok = FALSE)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1228  -0.1023  -0.0939  -0.0855   3.4939  
#> 
#> Coefficients:
#>                      Estimate Std. Error  z value Pr(>|z|)    
#> (Intercept)        -6.1718426  0.0408169 -151.208  < 2e-16 ***
#> assigned_treatment -0.1182655  0.0449731   -2.630  0.00855 ** 
#> for_period          0.0032507  0.0001481   21.952  < 2e-16 ***
#> followup_time       0.0016804  0.0002068    8.124 4.51e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 111015  on 1939052  degrees of freedom
#> Residual deviance: 110512  on 1939049  degrees of freedom
#> AIC: 110520
#> 
#> Number of Fisher Scoring iterations: 8
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   4.784   0.236   4.344 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                 names     estimate   robust_se            lb           ub
#> 1:        (Intercept) -6.171842626 0.347319209 -6.8525882755 -5.491096977
#> 2: assigned_treatment -0.118265496 0.293437150 -0.6934023091  0.456871318
#> 3:         for_period  0.003250720 0.001300476  0.0007017874  0.005799653
#> 4:      followup_time  0.001680353 0.001554794 -0.0013670440  0.004727750
#>              z  p_value
#> 1: -17.7699432  < 0.001
#> 2:  -0.4030352 0.686922
#> 3:   2.4996389 0.012432
#> 4:   1.0807557 0.279806
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period followup_time
#> (Intercept)         0.1206306329       9.612365e-03 -4.245403e-04 -3.600331e-04
#> assigned_treatment  0.0096123647       8.610536e-02 -7.143057e-05 -1.651184e-05
#> for_period         -0.0004245403      -7.143057e-05  1.691238e-06  1.070861e-06
#> followup_time      -0.0003600331      -1.651184e-05  1.070861e-06  2.417385e-06
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.988   0.181   1.168 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.004611686 0.009208130 0.013789351 0.018355370 0.022906168
#> 
#> $trt_1
#> [1] 0.004099572 0.008187863 0.012264880 0.016330628 0.020385084
#> 
# }