Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
trial_example$catvarA <- factor(trial_example$catvarA)
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = c("catvarA", "nvarA"),
  use_censor = 0,
  use_weight = 0,
  last_period = 10
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.653   0.048   0.701 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   4.860   0.804   5.667 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 110339"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[, weight])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1428  -0.1002  -0.0870  -0.0739   3.6290  
#> 
#> Coefficients:
#>                      Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -6.254e+00  2.637e-01 -23.715  < 2e-16 ***
#> assigned_treatment  1.330e-02  2.025e-01   0.066  0.94765    
#> for_period          3.892e-03  2.083e-03   1.869  0.06165 .  
#> I(for_period^2)    -6.159e-08  5.182e-06  -0.012  0.99052    
#> followup_time       5.314e-03  2.017e-03   2.635  0.00841 ** 
#> I(followup_time^2) -9.667e-06  7.360e-06  -1.313  0.18902    
#> catvarA1           -6.652e-01  1.447e-01  -4.598 4.28e-06 ***
#> catvarA2           -8.498e-01  1.991e-01  -4.269 1.97e-05 ***
#> catvarA3           -1.061e+01  2.194e+02  -0.048  0.96144    
#> catvarA7            8.603e-02  1.838e-01   0.468  0.63982    
#> nvarA              -9.058e-02  3.751e-02  -2.415  0.01574 *  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 5828.0  on 110338  degrees of freedom
#> Residual deviance: 5746.9  on 110328  degrees of freedom
#> AIC: 5768.9
#> 
#> Number of Fisher Scoring iterations: 14
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   0.811   0.204   1.028 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                  names      estimate    robust_se            lb            ub
#>  1:        (Intercept) -6.253757e+00 5.472757e-01 -7.326418e+00 -5.181097e+00
#>  2: assigned_treatment  1.329715e-02 4.123340e-01 -7.948775e-01  8.214718e-01
#>  3:         for_period  3.891814e-03 4.515678e-03 -4.958915e-03  1.274254e-02
#>  4:    I(for_period^2) -6.158817e-08 1.039899e-05 -2.044361e-05  2.032044e-05
#>  5:      followup_time  5.313692e-03 4.653611e-03 -3.807386e-03  1.443477e-02
#>  6: I(followup_time^2) -9.667148e-06 2.168391e-05 -5.216762e-05  3.283332e-05
#>  7:           catvarA1 -6.651683e-01 2.957883e-01 -1.244913e+00 -8.542326e-02
#>  8:           catvarA2 -8.498128e-01 3.927644e-01 -1.619631e+00 -7.999461e-02
#>  9:           catvarA3 -1.060516e+01 1.032819e+00 -1.262949e+01 -8.580837e+00
#> 10:           catvarA7  8.603119e-02 3.390479e-01 -5.785027e-01  7.505651e-01
#> 11:              nvarA -9.058413e-02 6.220449e-02 -2.125049e-01  3.133666e-02
#>                 z  p_value
#>  1: -11.427068755  < 0.001
#>  2:   0.032248490 0.974274
#>  3:   0.861844961 0.388773
#>  4:  -0.005922513 0.995275
#>  5:   1.141842750 0.253519
#>  6:  -0.445821224 0.655726
#>  7:  -2.248798685 0.024525
#>  8:  -2.163670721 0.030490
#>  9: -10.268172169  < 0.001
#> 10:   0.253743491 0.799694
#> 11:  -1.456231524 0.145329
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period
#> (Intercept)         2.995107e-01      -1.115652e-02 -2.079473e-03
#> assigned_treatment -1.115652e-02       1.700193e-01  2.422405e-04
#> for_period         -2.079473e-03       2.422405e-04  2.039135e-05
#> I(for_period^2)     3.625756e-06      -9.159570e-07 -4.377749e-08
#> followup_time      -1.929328e-04      -2.398614e-04 -5.817782e-06
#> I(followup_time^2) -1.535667e-06       7.467942e-07  3.766251e-08
#> catvarA1           -2.903288e-03       1.676501e-02  1.788376e-05
#> catvarA2           -3.915752e-02       1.318415e-02  3.262416e-04
#> catvarA3           -1.835278e-02       3.480597e-02  3.476442e-04
#> catvarA7           -8.498332e-02       1.667797e-02  5.321593e-04
#> nvarA               3.141628e-03      -3.297925e-03 -4.256128e-05
#>                    I(for_period^2) followup_time I(followup_time^2)
#> (Intercept)           3.625756e-06 -1.929328e-04      -1.535667e-06
#> assigned_treatment   -9.159570e-07 -2.398614e-04       7.467942e-07
#> for_period           -4.377749e-08 -5.817782e-06       3.766251e-08
#> I(for_period^2)       1.081390e-10  1.613033e-08      -8.489342e-11
#> followup_time         1.613033e-08  2.165610e-05      -9.174935e-08
#> I(followup_time^2)   -8.489342e-11 -9.174935e-08       4.701921e-10
#> catvarA1             -2.021093e-07 -2.020398e-04       6.305442e-07
#> catvarA2             -7.527130e-07 -4.327200e-04       2.235758e-06
#> catvarA3             -1.420683e-06 -8.514668e-05       1.882780e-07
#> catvarA7             -8.955145e-07 -4.600827e-05       2.430878e-07
#> nvarA                 6.884998e-08  6.354360e-05      -2.755293e-07
#>                         catvarA1      catvarA2      catvarA3      catvarA7
#> (Intercept)        -2.903288e-03 -3.915752e-02 -1.835278e-02 -8.498332e-02
#> assigned_treatment  1.676501e-02  1.318415e-02  3.480597e-02  1.667797e-02
#> for_period          1.788376e-05  3.262416e-04  3.476442e-04  5.321593e-04
#> I(for_period^2)    -2.021093e-07 -7.527130e-07 -1.420683e-06 -8.955145e-07
#> followup_time      -2.020398e-04 -4.327200e-04 -8.514668e-05 -4.600827e-05
#> I(followup_time^2)  6.305442e-07  2.235758e-06  1.882780e-07  2.430878e-07
#> catvarA1            8.749070e-02  2.280222e-02  7.580147e-02  1.485737e-02
#> catvarA2            2.280222e-02  1.542639e-01  4.607781e-02  2.690058e-02
#> catvarA3            7.580147e-02  4.607781e-02  1.066715e+00  1.746097e-02
#> catvarA7            1.485737e-02  2.690058e-02  1.746097e-02  1.149535e-01
#> nvarA              -3.068144e-03 -1.370007e-03 -5.359250e-03 -2.036404e-03
#>                            nvarA
#> (Intercept)         3.141628e-03
#> assigned_treatment -3.297925e-03
#> for_period         -4.256128e-05
#> I(for_period^2)     6.884998e-08
#> followup_time       6.354360e-05
#> I(followup_time^2) -2.755293e-07
#> catvarA1           -3.068144e-03
#> catvarA2           -1.370007e-03
#> catvarA3           -5.359250e-03
#> catvarA7           -2.036404e-03
#> nvarA               3.869398e-03
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.113   0.000   0.114 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.003596328 0.007195528 0.010797461 0.014401986 0.018008880
#> 
#> $trt_1
#> [1] 0.003644251 0.007291194 0.010940687 0.014592587 0.018246664
#>