Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
trial_example$catvarA <- factor(trial_example$catvarA)
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = c("catvarA", "nvarA"),
  use_censor = 0,
  use_weight = 0,
  last_period = 10
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.528   0.044   0.573 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   4.079   0.564   4.645 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 1939053"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[, weight])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1461  -0.1025  -0.0926  -0.0846   3.9179  
#> 
#> Coefficients:
#>                      Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -6.198e+00  7.168e-02 -86.473  < 2e-16 ***
#> assigned_treatment -9.363e-02  4.560e-02  -2.053  0.04006 *  
#> for_period          2.631e-03  6.007e-04   4.379 1.19e-05 ***
#> I(for_period^2)     3.071e-06  1.368e-06   2.245  0.02478 *  
#> followup_time       2.634e-03  5.235e-04   5.032 4.87e-07 ***
#> I(followup_time^2) -3.828e-06  2.291e-06  -1.670  0.09486 .  
#> catvarA1           -2.974e-02  2.672e-02  -1.113  0.26572    
#> catvarA2           -1.327e-01  3.783e-02  -3.509  0.00045 ***
#> catvarA3           -1.111e+01  3.486e+01  -0.319  0.74986    
#> catvarA7            1.944e-01  3.710e-02   5.240 1.60e-07 ***
#> nvarA              -6.865e-02  7.363e-03  -9.324  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 111015  on 1939052  degrees of freedom
#> Residual deviance: 110279  on 1939042  degrees of freedom
#> AIC: 110301
#> 
#> Number of Fisher Scoring iterations: 15
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>  16.689   1.128  17.868 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                  names      estimate    robust_se            lb            ub
#>  1:        (Intercept) -6.198159e+00 5.126578e-01 -7.202969e+00 -5.193350e+00
#>  2: assigned_treatment -9.363337e-02 2.795403e-01 -6.415323e-01  4.542656e-01
#>  3:         for_period  2.630526e-03 4.114880e-03 -5.434638e-03  1.069569e-02
#>  4:    I(for_period^2)  3.071113e-06 9.050305e-06 -1.466749e-05  2.080971e-05
#>  5:      followup_time  2.633784e-03 3.482277e-03 -4.191479e-03  9.459048e-03
#>  6: I(followup_time^2) -3.827510e-06 1.728384e-05 -3.770384e-05  3.004882e-05
#>  7:           catvarA1 -2.973889e-02 1.938786e-01 -4.097410e-01  3.502632e-01
#>  8:           catvarA2 -1.327333e-01 2.808822e-01 -6.832624e-01  4.177958e-01
#>  9:           catvarA3 -1.111350e+01 8.564779e-01 -1.279220e+01 -9.434808e+00
#> 10:           catvarA7  1.943945e-01 2.323162e-01 -2.609453e-01  6.497342e-01
#> 11:              nvarA -6.864810e-02 4.166260e-02 -1.503068e-01  1.301061e-02
#>               z  p_value
#>  1: -12.0902458  < 0.001
#>  2:  -0.3349548 0.737659
#>  3:   0.6392717 0.522646
#>  4:   0.3393380 0.734355
#>  5:   0.7563396 0.449446
#>  6:  -0.2214502 0.824742
#>  7:  -0.1533892 0.878091
#>  8:  -0.4725585 0.636528
#>  9: -12.9758210  < 0.001
#> 10:   0.8367667 0.402724
#> 11:  -1.6477150 0.099411
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period
#> (Intercept)         2.628181e-01      -1.613776e-03 -1.814193e-03
#> assigned_treatment -1.613776e-03       7.814277e-02  3.814160e-05
#> for_period         -1.814193e-03       3.814160e-05  1.693224e-05
#> I(for_period^2)     3.197132e-06      -2.499539e-07 -3.537578e-08
#> followup_time      -2.642055e-04      -5.470344e-05 -2.254897e-06
#> I(followup_time^2) -3.733070e-07       2.157575e-07  1.707582e-08
#> catvarA1           -1.559032e-02       3.827837e-03 -2.760991e-05
#> catvarA2           -2.373262e-02       4.705384e-03  5.182213e-05
#> catvarA3           -2.528093e-02       6.223373e-03  3.493958e-05
#> catvarA7           -4.630520e-02       1.873083e-03  2.428685e-04
#> nvarA              -4.893515e-04      -7.100799e-05  2.246996e-07
#>                    I(for_period^2) followup_time I(followup_time^2)
#> (Intercept)           3.197132e-06 -2.642055e-04      -3.733070e-07
#> assigned_treatment   -2.499539e-07 -5.470344e-05       2.157575e-07
#> for_period           -3.537578e-08 -2.254897e-06       1.707582e-08
#> I(for_period^2)       8.190802e-11  8.065190e-09      -4.183260e-11
#> followup_time         8.065190e-09  1.212626e-05      -5.419325e-08
#> I(followup_time^2)   -4.183260e-11 -5.419325e-08       2.987311e-10
#> catvarA1              8.637399e-08  3.700014e-05       4.847118e-09
#> catvarA2             -9.791245e-08  1.385078e-05       3.430456e-08
#> catvarA3              1.774293e-08 -4.499417e-05       3.966684e-07
#> catvarA7             -3.692330e-07 -3.912857e-05       2.737532e-07
#> nvarA                -2.000445e-09 -3.702054e-06      -1.510526e-08
#>                         catvarA1      catvarA2      catvarA3      catvarA7
#> (Intercept)        -1.559032e-02 -2.373262e-02 -2.528093e-02 -4.630520e-02
#> assigned_treatment  3.827837e-03  4.705384e-03  6.223373e-03  1.873083e-03
#> for_period         -2.760991e-05  5.182213e-05  3.493958e-05  2.428685e-04
#> I(for_period^2)     8.637399e-08 -9.791245e-08  1.774293e-08 -3.692330e-07
#> followup_time       3.700014e-05  1.385078e-05 -4.499417e-05 -3.912857e-05
#> I(followup_time^2)  4.847118e-09  3.430456e-08  3.966684e-07  2.737532e-07
#> catvarA1            3.758891e-02  2.376407e-02  1.997605e-02  1.408503e-02
#> catvarA2            2.376407e-02  7.889481e-02  2.350944e-02  1.729574e-02
#> catvarA3            1.997605e-02  2.350944e-02  7.335545e-01  1.568922e-02
#> catvarA7            1.408503e-02  1.729574e-02  1.568922e-02  5.397082e-02
#> nvarA              -3.696755e-04 -3.485475e-04 -4.843105e-04  1.368636e-04
#>                            nvarA
#> (Intercept)        -4.893515e-04
#> assigned_treatment -7.100799e-05
#> for_period          2.246996e-07
#> I(for_period^2)    -2.000445e-09
#> followup_time      -3.702054e-06
#> I(followup_time^2) -1.510526e-08
#> catvarA1           -3.696755e-04
#> catvarA2           -3.485475e-04
#> catvarA3           -4.843105e-04
#> catvarA7            1.368636e-04
#> nvarA               1.735773e-03
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   1.105   0.116   1.222 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.004537697 0.009064131 0.013579249 0.018082997 0.022575262
#> 
#> $trt_1
#> [1] 0.004133979 0.008259575 0.012376732 0.016485396 0.020585464
#>