Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = ~ nvarA + nvarB,
  include_followup_time_case = ~followup_time,
  include_expansion_time_case = ~for_period,
  last_followup = 30,
  last_period = 30,
  use_censor = 0,
  use_weight = 0
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.544   0.028   0.572 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   3.784   0.432   4.217 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 62981"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[["weight"]])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.7239  -0.0704  -0.0496  -0.0349   3.4643  
#> 
#> Coefficients: (2 not defined because of singularities)
#>                    Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -9.08925    0.86878 -10.462  < 2e-16 ***
#> assigned_treatment       NA         NA      NA       NA    
#> for_period          0.06736    0.02850   2.364  0.01810 *  
#> followup_time       0.08673    0.02859   3.034  0.00242 ** 
#> nvarA                    NA         NA      NA       NA    
#> nvarB               0.27146    0.04545   5.973 2.33e-09 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 339.02  on 10321  degrees of freedom
#> Residual deviance: 297.34  on 10318  degrees of freedom
#> AIC: 305.34
#> 
#> Number of Fisher Scoring iterations: 10
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   0.038   0.000   0.055 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.047   0.000   0.047 

predict_survival(i, predict_times = seq(1, 20, by = 2))
#> Warning: prediction from a rank-deficient fit may be misleading
#> Warning: prediction from a rank-deficient fit may be misleading
#> $trt_0
#>  [1] 0.001581042 0.003267886 0.005052393 0.006928112 0.008892979 0.010952155
#>  [7] 0.013120567 0.015424805 0.017904216 0.020585237
#> 
#> $trt_1
#>  [1] 0.001581042 0.003267886 0.005052393 0.006928112 0.008892979 0.010952155
#>  [7] 0.013120567 0.015424805 0.017904216 0.020585237
#>