Predict Cumulative Incidence

Predict Cumulative Incidence

predict_survival(object, model, predict_times, newdata)

Arguments

object

The result from initiators. Either object or model must be specified.

model

A glm object.

predict_times

a vector of follow up times to predict values for.

newdata

data.frame to predict survival for. Extracted from model/object if not provided.

Value

a list with estimated cumulative incidence curves for assigned_treatment

equal to 0 and 1.

Examples

data("trial_example")
trial_example$catvarA <- factor(trial_example$catvarA)
i <- initiators(
  data = trial_example,
  id = "id",
  period = "period",
  eligible = "eligible",
  treatment = "treatment",
  outcome = "outcome",
  outcome_cov = c("catvarA", "nvarA"),
  use_censor = 0,
  use_weight = 0,
  last_period = 10
)
#> [1] "Start data manipulation"
#> [1] "Finish data manipulation"
#> [1] "Processing time of data manipulation:"
#>    user  system elapsed 
#>   0.545   0.028   0.574 
#> [1] "----------------------------"
#> [1] "Start data extension"
#> [1] "Finish data extension"
#> [1] "Processing time of data extension:"
#>    user  system elapsed 
#>   4.369   0.560   4.932 
#> [1] "----------------------------"
#> [1] "Number of observations in expanded data: 96873"
#> [1] "Base Analysis"
#> 
#> Call:
#> stats::glm(formula = model_formula, family = binomial(link = "logit"), 
#>     data = data, weights = data[, weight])
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -0.1327  -0.1029  -0.0909  -0.0764   3.5825  
#> 
#> Coefficients:
#>                      Estimate Std. Error z value Pr(>|z|)    
#> (Intercept)        -6.250e+00  2.629e-01 -23.775  < 2e-16 ***
#> assigned_treatment -1.168e-01  2.348e-01  -0.498  0.61882    
#> for_period          6.148e-03  2.033e-03   3.025  0.00249 ** 
#> I(for_period^2)    -6.641e-06  4.972e-06  -1.336  0.18168    
#> followup_time       1.300e-03  2.086e-03   0.623  0.53313    
#> I(followup_time^2) -2.092e-06  7.894e-06  -0.265  0.79103    
#> catvarA1           -2.954e-01  1.361e-01  -2.170  0.03000 *  
#> catvarA2           -4.039e-01  1.978e-01  -2.042  0.04115 *  
#> catvarA3           -1.124e+01  2.309e+02  -0.049  0.96117    
#> catvarA7            1.911e-01  1.787e-01   1.070  0.28480    
#> nvarA              -4.723e-02  3.350e-02  -1.410  0.15852    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for binomial family taken to be 1)
#> 
#>     Null deviance: 5343.2  on 96872  degrees of freedom
#> Residual deviance: 5296.8  on 96862  degrees of freedom
#> AIC: 5318.8
#> 
#> Number of Fisher Scoring iterations: 15
#> 
#> [1] "-------------------------------------------------"
#> [1] "Processing time of modeling for base analysis:"
#>    user  system elapsed 
#>   0.806   0.143   0.972 
#> [1] "Base Analysis with robust variance"
#> [1] "-------------------------------------------------------"
#> [1] "Robust standard error:"
#> $summary
#>                  names      estimate    robust_se            lb            ub
#>  1:        (Intercept) -6.249931e+00 5.438470e-01 -7.315871e+00 -5.183991e+00
#>  2: assigned_treatment -1.168273e-01 3.233118e-01 -7.505185e-01  5.168638e-01
#>  3:         for_period  6.148405e-03 4.274241e-03 -2.229107e-03  1.452592e-02
#>  4:    I(for_period^2) -6.640959e-06 9.930096e-06 -2.610395e-05  1.282203e-05
#>  5:      followup_time  1.299838e-03 3.944021e-03 -6.430443e-03  9.030119e-03
#>  6: I(followup_time^2) -2.091638e-06 1.819280e-05 -3.774953e-05  3.356625e-05
#>  7:           catvarA1 -2.954181e-01 2.540286e-01 -7.933141e-01  2.024780e-01
#>  8:           catvarA2 -4.039127e-01 3.434556e-01 -1.077086e+00  2.692604e-01
#>  9:           catvarA3 -1.124102e+01 1.022465e+00 -1.324505e+01 -9.236989e+00
#> 10:           catvarA7  1.911109e-01 2.950330e-01 -3.871537e-01  7.693755e-01
#> 11:              nvarA -4.723496e-02 4.057838e-02 -1.267686e-01  3.229867e-02
#>               z p_value
#>  1: -11.4920751 < 0.001
#>  2:  -0.3613457 0.71784
#>  3:   1.4384788 0.15030
#>  4:  -0.6687708 0.50364
#>  5:   0.3295718 0.74172
#>  6:  -0.1149706 0.90847
#>  7:  -1.1629322 0.24486
#>  8:  -1.1760258 0.23958
#>  9: -10.9940359 < 0.001
#> 10:   0.6477611 0.51714
#> 11:  -1.1640424 0.24441
#> 
#> $matrix
#>                      (Intercept) assigned_treatment    for_period
#> (Intercept)         2.957696e-01      -1.451627e-02 -1.756399e-03
#> assigned_treatment -1.451627e-02       1.045305e-01  8.953411e-05
#> for_period         -1.756399e-03       8.953411e-05  1.826914e-05
#> I(for_period^2)     2.846568e-06      -4.085518e-07 -3.980041e-08
#> followup_time      -3.144948e-04       3.042872e-05 -4.440465e-06
#> I(followup_time^2) -7.418704e-07      -3.444252e-08  2.650637e-08
#> catvarA1           -3.069319e-02       9.498421e-03 -4.888511e-05
#> catvarA2           -4.413154e-02       9.879003e-03 -4.349860e-05
#> catvarA3           -4.752427e-02       1.092469e-02 -1.194016e-04
#> catvarA7           -7.660428e-02       1.046230e-02  2.208520e-04
#> nvarA              -6.551603e-04      -5.133775e-04 -4.181308e-06
#>                    I(for_period^2) followup_time I(followup_time^2)
#> (Intercept)           2.846568e-06 -3.144948e-04      -7.418704e-07
#> assigned_treatment   -4.085518e-07  3.042872e-05      -3.444252e-08
#> for_period           -3.980041e-08 -4.440465e-06       2.650637e-08
#> I(for_period^2)       9.860681e-11  1.346185e-08      -6.280418e-11
#> followup_time         1.346185e-08  1.555530e-05      -6.199295e-08
#> I(followup_time^2)   -6.280418e-11 -6.199295e-08       3.309780e-10
#> catvarA1              2.440001e-07  6.459680e-05      -1.230801e-07
#> catvarA2              3.355554e-07  7.228618e-05      -5.632555e-08
#> catvarA3              6.434908e-07  8.705819e-05       5.904117e-08
#> catvarA7             -7.760318e-08 -4.982624e-05       4.649048e-07
#> nvarA                 1.168257e-08 -1.571078e-05       4.344731e-09
#>                         catvarA1      catvarA2      catvarA3      catvarA7
#> (Intercept)        -3.069319e-02 -4.413154e-02 -4.752427e-02 -7.660428e-02
#> assigned_treatment  9.498421e-03  9.879003e-03  1.092469e-02  1.046230e-02
#> for_period         -4.888511e-05 -4.349860e-05 -1.194016e-04  2.208520e-04
#> I(for_period^2)     2.440001e-07  3.355554e-07  6.434908e-07 -7.760318e-08
#> followup_time       6.459680e-05  7.228618e-05  8.705819e-05 -4.982624e-05
#> I(followup_time^2) -1.230801e-07 -5.632555e-08  5.904117e-08  4.649048e-07
#> catvarA1            6.453054e-02  3.321439e-02  2.844323e-02  2.789829e-02
#> catvarA2            3.321439e-02  1.179618e-01  3.575394e-02  3.621426e-02
#> catvarA3            2.844323e-02  3.575394e-02  1.045436e+00  3.820090e-02
#> catvarA7            2.789829e-02  3.621426e-02  3.820090e-02  8.704444e-02
#> nvarA              -9.503278e-05  1.219878e-03  2.220492e-04  1.085530e-03
#>                            nvarA
#> (Intercept)        -6.551603e-04
#> assigned_treatment -5.133775e-04
#> for_period         -4.181308e-06
#> I(for_period^2)     1.168257e-08
#> followup_time      -1.571078e-05
#> I(followup_time^2)  4.344731e-09
#> catvarA1           -9.503278e-05
#> catvarA2            1.219878e-03
#> catvarA3            2.220492e-04
#> catvarA7            1.085530e-03
#> nvarA               1.646605e-03
#> 
#> [1] "----------------------------------------------"
#> [1] "Processing time of getting the output and sandwich with reduced switch data:"
#>    user  system elapsed 
#>   0.098   0.000   0.098 

predict_survival(i, predict_times = c(1, 2, 3, 4, 5))
#> $trt_0
#> [1] 0.004653340 0.009288734 0.013906210 0.018505799 0.023087500
#> 
#> $trt_1
#> [1] 0.004142604 0.008271572 0.012386914 0.016488641 0.020576739
#>