Weight Calculation Function

This function performs the calculation for weight of the data

weight_func(
  sw_data,
  cov_switchn = NA,
  model_switchn = NA,
  class_switchn = NA,
  cov_switchd = NA,
  model_switchd = NA,
  class_switchd = NA,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  cense = NA,
  pool_cense = 0,
  cov_censed = NA,
  model_censed = NA,
  class_censed = NA,
  cov_censen = NA,
  model_censen = NA,
  class_censen = NA,
  include_regime_length = 0,
  numCores = NA,
  save_dir,
  quiet = FALSE
)

Arguments

sw_data

A data.table

cov_switchn

List of covariates to be used in logistic model for switching probabilities for numerator model

model_switchn

List of models (functions) to use the covariates from cov_switchn

class_switchn

Class variables used in logistic model for nominator model

cov_switchd

List of covariates to be used in logistic model for switching probabilities for denominator model

model_switchd

List of models (functions) to use the covariates from cov_switchd

class_switchd

Class variables used in logistic model for denominator model

eligible_wts_0

Eligibility criteria used in weights for model condition Am1 = 0

eligible_wts_1

Eligibility criteria used in weights for model condition Am1 = 1

cense

Censoring variable

pool_cense

Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model) Defaults to 0

cov_censed

List of covariates to be used in logistic model for censoring weights in denominator model

model_censed

List of models (functions) to use the covariates from cov_censed

class_censed

Class variables used in censoring logistic regression in denominator model

cov_censen

List of covariates to be used in logistic model for censoring weights in numerator model

model_censen

List of models (functions) to use the covariates from cov_censen

class_censen

Class variables used in censoring logistic regression in numerator model

include_regime_length

If defined as 1 a new variable (time_on_regime) is added to dataset - This variable stores the duration of time that the patient has been on the current treatment value

numCores

Number of cores for parallel programming

save_dir

Directory to save tidy weight model summaries in as 'weight_models.rda'

quiet

Don't print progress messages.