Weight Calculation Function

This function performs the calculation for weight of the data

weight_func(
  sw_data,
  switch_n_cov = NA,
  switch_d_cov = NA,
  eligible_wts_0 = NA,
  eligible_wts_1 = NA,
  cense = NA,
  pool_cense = 0,
  cense_d_cov = NA,
  cense_n_cov = NA,
  include_regime_length = 0,
  save_weight_models = FALSE,
  save_dir,
  quiet = FALSE
)

Arguments

sw_data

A data.table

switch_n_cov

A RHS formula for modelling probability of switching treatment. Used in the numerator of weight calculation.

switch_d_cov

A RHS formula for modelling probability of switching treatment. Used in the denominator of weight calculation.

eligible_wts_0

Eligibility criteria used in weights for model condition Am1 = 0

eligible_wts_1

Eligibility criteria used in weights for model condition Am1 = 1

cense

Censoring variable

pool_cense

Pool the numerator and denominator models (0: split models by previous treatment Am1 = 0 and Am1 = 1 as in treatment models and 1: pool all observations together into a single numerator and denominator model) Defaults to 0

cense_d_cov

A RHS formula for modelling probability of being censored. Used in the numerator of weight calculation.

cense_n_cov

A RHS formula for modelling probability of being censored. Used in the denominator of weight calculation.

include_regime_length

If defined as 1 a new variable (time_on_regime) is added to dataset. This variable stores the duration of time that the patient has been on the current treatment value

save_weight_models

Save weight models objects in data_dir.

save_dir

Directory to save tidy weight model summaries in as 'weight_models.rda'

quiet

Don't print progress messages.